Update prometheus.md
Customer brought up that the syntax in our documentation for Docker is incorrect. 

The previous documentation listed metrics as an array of strings with colon pairs. However, they received the following error

```
2020-02-26 15:01:11 UTC | CORE | DEBUG | (pkg/collector/python/datadog_agent.go:122 in LogMessage) | openmetrics:burrow:cb2825231e8b6144 | (mixins.py:552) | Skipping metric `kafka_burrow_partition_status` as it is not defined in the metrics mapper, has no transformer function, nor does it match any wildcards.
```

When they changed the metrics to an array of objects, the integration worked.

The previous format: 
\"metrics\":[\"<PROMETHEUS_METRIC_TO_FETCH>: <DATADOG_NEW_METRIC_NAME>\"]

Instead it should be:
 \"metrics\":[{\"<PROMETHEUS_METRIC_TO_FETCH>\": \"<DATADOG_NEW_METRIC_NAME>\"}]

The format is correct on the kubernetes tab of the documentation